Fix dependency problem with xinetd client/server
Previously, the xinetd services all relied either on Package[amanda] or
Package[amanda/client]. This did not account for situations where the amanda
client was not being installed; e.g. a server-only install.

This commit changes the "require" of xinetd services to $post_package;
a variable which all amanda packages include as a "before" parameter.