Massively refactor module, redo how configs work
This commit refactors most or all aspects of the amanda module. It still
resembles the 0.0.3 version, but significant structural and interface
changes have been made which renders it completely
non-backwards-compatible.

The objective of this refactor is to clean up and standardize the code,
and to make it make more sense. Towards this end most variables were
renamed and how configs work was redone from close to scratch.