Switch Debian params from zmanda packages to dist
Previously, the params class referenced the Zmanda packages for
Ubuntu/Debian, rather than the distribution provided packages. This
commit changes the params class to specify the distribution packages in
order to make the module more portable to other organizations
(internally we built/backported the versions we needed in order to make
this change work for us too).