Set type UNLISTED for amandaidx and amidxtape in xinetd
The entries may be present in some /etc/services files but
we can't guarantee their inclusion given they've never been
formally registered with IANA [1]. Therefore setting type
to UNLISTED in the xinetd configs for these services is
appropriate and the only way to guarantee the services work
properly. The addition of this config value will not have
a negative effect on systems which contain the services
entries.

[1] http://www.iana.org/assignments/service-names-port-numbers/service-names-port-numbers.txt

closes #17